Added basic support for dependencies for tasks.
Dependencies are containers that have to be started before the task container
can run.

If the dependency container has a healthcheck, we wait until the container is
considered healthy before starting the task container.

Dependencies are started one-by-one (ie. not in parallel - the first dependency
must be healthy before the next one is started).

Dependencies are stopped when the task container stops. There is currently no
handling in place for the situation where a dependency exits before the task
container stops.

All containers (any dependencies and task container) are run in an isolated
network can can refer to each other by their name (ie. if a container has the
name 'http-server', resolving 'http-server' returns the IP address of that
container).