Upgrade to Gradle 4.0.2.
This involved switching to the new plugins DSL in build.gradle for the Kotlin
plugin (otherwise it would produce lots of errors). Unfortunately the new DSL
does not support using variables to define the version, so we have to specify it
explicitly in both the plugins block and later again in the dependencies block.