Split EventLogger into an interface and its current implementation.
This is to make way for the "non-simple" event logger that uses fancier console
effects (eg. replacing text) that isn't appropriate for all situations - we'll
still use the simple event logger in situations where that is desirable (eg. on
CI or when the output is being piped to a file rather than displayed to a user).