Refactor ApplicationRunner used in journey tests and make assertions stronger.
Previously we were looking for some output that was also included in the
command line printed before the command was run - so we weren't actually
checking that the output was shown at all. So we check that the output contains
a line ending with the expected output (the command is shown as part of a longer
line).

However, for some reason, the line with the output ends with both a carriage
return (\r) and line feed (\n) character, so we have to look for a substring
in the output with both of those characters at the end.