Don't try to use the fancy event logger if the terminal doesn't support it.
On Travis, we get output like this:
https://travis-ci.org/charleskorn/batect-sample/jobs/278033671#L511

This is because the simple event logger is not enabled when it should be.
For now, just detect when TERM has been set to 'dumb' (or not set at all)
and switch to the simple event logger. Ideally, we'd use something like
the terminfo database to find the terminal's capabilities, but this is a
simple solution for the time being.