Add logging basic application and system information on startup.
This happens after we've passed common command line options (including
the log path), but before we start parsing command-specific options so
that logging is available as early as possible.

Also changed the Docker client to catch and deal with exceptions that
arise while getting client and server version information, as nothing
needs to know when this fails (ie. nothing depends on that method
throwing an exception) and the behaviour of turning the error into a
string for display to a user is now needed in two places ('version'
command and when logging system info on startup).