Force simple output mode on Travis.
Travis' ANSI escape code handling is patchy and doesn't support things
like cursor movement or clearing lines with \r, yet reports itself as
TERM=xterm, so we default to using the fancy output logger. This results
in output like
https://travis-ci.org/charleskorn/batect-sample-ruby/jobs/301149186.

There's no easy way to detect this situation - where TERM suggests the
terminal should be able to handle these escape codes but it doesn't - so
we have to add a special case for Travis.

To further complicate matters, Travis sets TERM=dumb for Java projects
but not for other languages, so this problem only showed up in the Ruby
sample project.