Add the ability to override health check options in the Dockerfile.
The interval, number of retries and start period can all be overridden
in batect.yml (the command itself or the command timeout can't be).

This is useful if you want to use an image in production with health
check options baked into the image, but want to specify different
options during development (eg. to decrease the interval and number of
retries because you expect the service to start far quicker in a test
scenario).