Fix the issue where the fancy output mode output would become garbled.
This happens when we try to print a line of text longer than the width
of the console and then move the cursor up to replace the text, but this
would put the cursor part-way through the (now wrapped) line of text.

Now we limit the length of lines printed to the width of the console.

Note that if the user resizes the console after we retrieve the console
size, we won't recognise this and so the output will either be too short
or too long.