Add the ability to run a container with the current UID and GID.
This means that any files created in mounted directories are owned by
the executing user, rather than root. This doesn't affect OS X, but does
affect Linux, and happens because the Docker daemon runs as root on
Linux. (On OS X, Docker for Mac runs as the executing user and so files
are created as that user.)

To do this, we start the container with --user <UID>:<GID> and mount new
/etc/passwd and /etc/group files into the container. Note that this
means that all users and groups apart from the executing user and root
are removed. (A future enhancement would be to pull the existing files
out of the image, add the extra line for the executing user and then
mount that, but that's far more complicated and we'd also have to deal
with potential conflicts as well.)

Still remaining to do:
* update documentation to include this feature
* make sure any directories that are mounted into the container are
  created before the container is started (otherwise Docker
  automatically creates them as root before starting the container)