Fix issue where the wrapper would fail if another invocation was running.
This happened because we would test to see if the JAR existed in the
expected location, but curl creates the file as soon as it starts
downloading - not after the download has completed successfully - and so
we tried to start the JAR when it was only partially downloaded.

Unfortunately this isn't easy to reliably test, so there isn't a test
case for this.