Fix the issue where running a built image can fail if the build output contains certain phrases.
We look for a line matching 'Successfully built XXX', but there's
nothing stopping a command that runs during the image build process from
printing something that looks like that. pip, for example, prints this
after it's compiled a module. Previously, we were finding the first
matching line and using that value as the image ID, which obviously
won't work if that first line isn't the line from Docker.