Create the user's home directory in 'run as current user' mode.
We can't do something like 'docker exec <container> mkdir /home/the-user'
because that runs as the container's user, which doesn't have permission
to create directories in /home (or a bunch of other places).

Mounting a directory would solve the problem, but this is unnecessary
and introduces a performance penalty for reads and writes, so instead we
create an empty directory locally with the correct owner and copy it
into the container.

Ideally we'd also use this technique for the /etc/passwd and /etc/group
files as well.