Fix the issue where upgrading batect could result in weird errors.
This happens if the upgrade adds or removes lines from the wrapper
script, as Bash does not read the entire file into memory before
running it and so continues execution from the same place in the (now
modified) script after batect exits.

With this change, Bash is forced to read the entire script into memory
immediately, preventing weird things from happening.

This won't fix the issue for upgrades to this version, but will fix it
for subsequent upgrades.