Add a tool to display the logs in Kibana.
This identified an issue in TaskRunner where we used the key 'task' to
refer to the task name, where in other places, the key 'task' was used
to refer to the whole task object. Elasticsearch didn't like this (it
wants the structure of data associated with keys to be consistent), so
we should be consistent too.