Fix the issue where the home directory was sometimes not owned by the desired user.
This would be triggered if both of the following were true:
- run_as_current_user was enabled
- a volume mount was configured to mount inside the container beneath
  the home directory (directly or indirectly)

This is because the volume mount is created first, and so all
directories leading up to the target directory are created, including
the home directory. Then, when we were running 'docker cp' to copy the
home directory into the container, it would be copied into the existing
home directory, rather than replacing it - leading to the home directory
still being owned by root and the home directory being polluted with a
random 'batect-home-(random number)' directory.

So, instead, we just always mount the home directory as a volume mount
instead of trying to copy it in.

For simplicity, we also revert to the old behaviour of mounting in the
/etc/passwd and /etc/group files as well (which reverts commit 30987ac).