Add a basic schema for the configuration file.
The purpose of this is to enable autocompletion and validation in IDEs,
not to serve as a definitive source of truth for valid configuration
files. (The only way to be sure is to try to run batect with the file.)

This enables editors like VS Code (when running with the 'YAML Support
by Red Hat' plugin) and JetBrains IDEA-bassed IDEs such as IntelliJ or
RubyMine (version 2018.2 or above) to provide autocompletion and
validation when editing configuration files.