Add workaround for #41.
The integration tests occasionally fail due to
https://github.com/square/okhttp/issues/2738. This appears to only
affect the integration tests on OS X, and only occurs occasionally.

The failure is a false negative (it's an issue in OkHttp or a
compatibility issue between OkHttp and the server in one of the test
images we use), so it's safe to retry a few times and declare the test
passing if it passes once.