Fix flaky unit test on Windows.
There is a race condition where the client will send the data to the
server before the server has started, causing the data to be dropped
and the server to receive a ERROR_NO_DATA error when calling
ConnectNamedPipe. Waiting for the server to connect and send some data
is enough to ensure that the data sent by the client is not lost.