Don't use progress events to trigger new work.
Image pull progress and image build progress events are just used so
that we can give the user feedback on long-running operations. They
can't trigger new work. Previously, we were sending them to the state
machine, triggering a recalculation and storing them in the event
history for later use as well. These are all unnecessary and expensive
operations (some of them involve locks that block looking for other
work) and result in slower build times and higher peak memory usage,
especially when running a task for the first time that requires building
or pulling a large number of large or complex images.