Stop building an image immediately if the user presses Ctrl-C.
Previously, we show the message acknowledging that the Ctrl-C had been
received but continue waiting for the build to finish. Now, we cancel
the build altogether.

Note that we won't implement this behaviour for most other operations
(eg. creating a container) because otherwise there would be a race
condition: if we cancelled a request just after the daemon had say
created the container but before we'd received its response, then we
wouldn't know that the container had been created and therefore wouldn't
know to clean it up.