Don't allow running batect on a 32-bit JRE on Windows.
There's something not quite right in our Win32 API calls that causes
them to hang if the JRE is 32-bit. While fixing that is the long-term
solution, that's not a simple thing to debug and this at least directs
users to something that works for the time being.

Given that Docker only runs on 64-bit machines, expecting a 64-bit JVM
is not a huge imposition.