Fix the issue where stdin was not connected to containers after a prerequisite task had run.
In addition to preventing input to these containers, it also prevented
containers from obtaining the current cursor position, which led to
garbled output in some situations.

A common example was running 'dotnet test' after a prerequisite that
built the application - the test summary would appear at the top of
the console instead of at the end of the output. This also caused
'dotnet test' to hang for 10 seconds while it tried twice to
(unsuccessfully) get the current cursor position.

This was caused by the stdin streaming for the previous task not
stopping correctly. Stopping a read from stdin in Java is non-trivial,
https://hypirion.com/musings/how-to-cancel-a-blocking-read has more
details on the solution.