Move Docker client into its own module.
This paves the way for moving to Kotlin/Native in a more incremental
fashion and improves build and test times locally (as unmodified
modules don't need to be rebuilt or tested).

Many of the tests are still in the wrong module - trying to keep this
commit as small as possible.