Rejig telemetry configuration.
This allows us to preserve the user ID across sessions even if the
user hasn't opted-in to telemetry on a full-time basis.

This is important for CI environments where the telemetry consent is
provided by an environment variable, and so the consent on disk is
always 'none' - otherwise we would get a new user ID for every
session, which isn't quite right and would skew the data.

To preserve the ability to completely disassociate with the previous
user ID, when --permanently-disable-telemetry is passed, the user ID
is regenerated.