Fix the issue where the Fish completion script would show a "no checksum lines found" error if the configuration file was invalid.
This was causing occasional build flakiness, as we'd occasionally
end up running a test in a directory that wasn't a valid project name
(eg. it ended with an underscore). We now make sure that we run the
tests from a temporary directory that is guaranteed to be a valid
project name.

https://github.com/batect/batect/pull/726/checks?check_run_id=1388970534
is one example of a flaky build.