Fix issue where FileSyncService could deadlock on machines with less than three physical CPU cores.
Dispatchers.Default limits the number of concurrently executing
coroutines to the number of physical CPU cores. We have a number of
blocking request.read() and response.write() calls in the three
coroutines we launch, so it's easy to consume this limit.

Furthermore, Dispatchers.Default is shared globally, so if we have
multiple concurrent FileSyncServices running (eg. to service different
image builds), then we're even more likely to consume this limit and
deadlock.

https://elizarov.medium.com/blocking-threads-suspending-coroutines-d33e11bf4761
explains this further.