Add `-h` as an alias for `--help`
Users typically expect to run `$foo_program -h` for help.  The GNU
standard on command-line options only mentions `--help` and `--version`;
a random survey of typical programs show the large majority using `-h`
as an alias for `--help` (OS-provided, or OSS). There are some legacy
programs (going back to the 1970s and 1980s) treating `-h` otherwise,
but the are a strong minority.