Gradle: Migrate `versioning` from Groovy to Kotlin script
This also swaps out `git` CLI runs with JGit (and thus gets rid of
`utilities.gradle`) to fix building Batect on Windows, which previously
might have resulted in

    A problem occurred evaluating script.
    > Cannot run program "git": CreateProcess error=2, The system cannot find the file specified

if `git.exe` was not added to the `PATH` during the Git for Windows
installation.

Note that the previously used `--candidates=0` is synonym with
`--exact-match`. Neither exists directly in JGit, but this can be
accounted for by manually disallowing tag matches that include the "-g"
suffix, also see [1].

Finally, move the `reckon` configuration to where the plugin is applied,
as with Kotlin script this otherwise results in a compile error.

[1]: https://git-scm.com/docs/git-describe#_examples